Separate sample module definitions from declarations
